ref(prettier): move tslint rule into own directory
