+++ src/prettier/rules/prettier/register-prettier.ts
@@ -1,5 +1,5 @@
 import { Rule, Tree } from '@angular-devkit/schematics';
-import { PackageJson } from '../lib';
+import { PackageJson } from '../../../lib';
 export function registerPrettier(): Rule {
   return (tree: Tree) => {
     const packageJson = new PackageJson(tree.read('package.json'));
