+++ src/prettier/rules/tslint/index.ts
