+++ src/prettier/rules/tslint/patch-ts-lint-configuration.spec.ts
